# reduce the # of allocations using pools